Relocate model definitions so same can be reused for daily and block batch models
