--- src/op_analytics/datapipeline/etl/intermediate/markers.py
@@ -1,1 +1,0 @@
-INTERMEDIATE_MODELS_MARKERS_TABLE = "intermediate_model_markers"
