+++ src/op_analytics/datapipeline/models/compute/querybuilder.py
@@ -31,7 +31,7 @@ class TemplatedSQLQuery:
 
     def render(self) -> RenderedSQLQuery:
         # Load templates from the templates folder.
-        tempates_dir = os.path.join(os.path.dirname(__file__), "templates")
+        tempates_dir = os.path.join(os.path.dirname(__file__), "../templates")
         env = Environment(loader=FileSystemLoader(tempates_dir))
 
         # Load and render the template
