+++ src/op_analytics/datapipeline/models/templates/contract_creation_traces.sql.j2
