+++ src/op_analytics/datapipeline/models/templates/daily_address_summary.sql.j2
