+++ tests/op_datasets/etl/intermediate/models/test_contract_creation_traces.py
@@ -1,7 +1,7 @@
 from datetime import date
 
 from op_analytics.coreutils.testutils.inputdata import InputTestData
-from op_analytics.datapipeline.etl.intermediate.testutils import IntermediateModelTestBase
+from op_analytics.datapipeline.models.compute.testutils import IntermediateModelTestBase
 
 CONTRACT_ADDRESS = "0xa16b2bc8053a320620a1ef33da325491ec064e4b"
