+++ tests/op_datasets/etl/intermediate/models/test_daily_address_summary.py
@@ -2,7 +2,7 @@ from datetime import date
 from decimal import Decimal
 
 from op_analytics.coreutils.testutils.inputdata import InputTestData
-from op_analytics.datapipeline.etl.intermediate.testutils import IntermediateModelTestBase
+from op_analytics.datapipeline.models.compute.testutils import IntermediateModelTestBase
 
 
 SYSTEM_ADDRESS = "0xdeaddeaddeaddeaddeaddeaddeaddeaddead0001"
